Update MainActivity
I added in the Holiday blocks. Basically when the events for a day are being presented. It will check the month and day to see if they are holidays. IF they are they will be added to a temporary array that will then be added to the bigger array that is being returned.